% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaytimes.R
\name{compute.time.range}
\alias{compute.time.range}
\title{Compute the time range from the block parameters}
\usage{
compute.time.range(block.parms)
}
\arguments{
\item{block.parms}{the data frame containing the geometric information 
about the time blocks to be plotted}
}
\value{
a vector of (min time , max time)
}
\description{
Compute the time range from the block parameters
}
