% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaytimes.R
\name{get.header.names}
\alias{get.header.names}
\title{Get the names in the header column values from a time table data frame}
\usage{
get.header.names(time.table, group.by = "day.of.week")
}
\arguments{
\item{time.table}{a data frame with columns start.time, end.time, day.of.week}

\item{group.by}{name of the grouping variable column in time.table data frame}
}
\value{
vector of unique column names
}
\description{
Get the names in the header column values from a time table data frame
}
