% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaytimes.R
\name{get.marked.times}
\alias{get.marked.times}
\title{Determine the times at which a line marking should be drawn}
\usage{
get.marked.times(min.max.time, n.marks = NULL)
}
\arguments{
\item{min.max.time}{a vector of (min time, max time)}

\item{n.marks}{number of marks to produce. If a non-integer is provided
marks are made at each hour}
}
\value{
vector of times
}
\description{
Determine the times at which a line marking should be drawn
}
