% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaytimes.R
\name{make.blocks}
\alias{make.blocks}
\title{Draw the time blocks according to block.parms}
\usage{
make.blocks(block.parms)
}
\arguments{
\item{block.params}{data frame of time block information}
}
\description{
Draw the time blocks according to block.parms
}
