% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaytimes.R
\name{make.hour.lines}
\alias{make.hour.lines}
\title{Mark the hour lines on the main plot area}
\usage{
make.hour.lines(min.max.time, n.marks, alpha = 0.5)
}
\arguments{
\item{min.max.time}{vector of (min time, max time)}

\item{n.marks}{the number of time marking to draw}

\item{alpha}{transparency of the marked lines}
}
\description{
Mark the hour lines on the main plot area
}
