% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaytimes.R
\name{make.time.scale}
\alias{make.time.scale}
\title{Draw the time scale on the plot}
\usage{
make.time.scale(min.max.time, n.marks, alpha = 0.5)
}
\arguments{
\item{min.max.time}{vector of (min time, max time)}

\item{n.marks}{number of markings to make}

\item{alpha}{transparency for the time marking names}
}
\description{
Draw the time scale on the plot
}
